server/talent_agent_backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
transformers==4.50.0
torch==2.6.0
requests==2.32.3
coreapi==2.3.3
PyYAML==6.0.2
